todo-cli/pyproject.toml
[tool.poetry]
name = "todo-cli"
version = "0.1.0"
description = "basic to-do list CLI"
authors = ["Tristan Sargent"]
readme = "README.md"

[tool.poetry.scripts]


[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



